requirements.txt
pandas>=2.0.0
dask[complete]>=2023.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
scikit-learn>=1.3.0
xgboost>=2.0.0
streamlit>=1.30.0
plotly>=5.0.0
python-dotenv>=1.0.0
jupyterlab>=4.0.0
numpy>=1.24.0
pytest>=7.0.0
fpdf>=1.7.2  # For report generation if needed
openai>=1.0.0 # Placeholder for LLM integration
